+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Reference documentation in `package.json`
 - Make using namespaces optional
 - Configure using directives
+- Remove `IsResource` property from `IResourceData`. Scene path generation now depends on file extension
 
 # 0.1.0
 - Add initial project